pyproject.toml
[project]
name = "azure-docintel-extract"
version = "0.1.0"
description = "Extract structured data from documents using Azure Document Intelligence"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-formrecognizer>=3.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "rich>=13.7.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.12.1",
    "ruff>=0.1.9"
]
